Revert "dracut: Randomize disk GUID only after ignition-disks succeeds"
We're going to use a flag file in the ESP, and not the GUID, to indicate
that this is a first boot. We'll still need to randomize the disk GUID so
each installation has a different one, but that can be done independently of
any Ignition run.

This reverts commit 7eb320ad7e6df729c4337c84b5f47e178c2e874b.

(cherry picked from commit a1ebd91955008f4cb6124bc16884ec10c0bdc863)