dracut/35torcx: Reimplement BindReadOnlyPaths to drop mount points
The BindReadOnlyPaths directive will create an empty file for the
mount point if it does not exist, and it remains in place after the
service stops.  In this case, an empty /etc/resolv.conf causes
tmpfiles to skip creating the symlink to resolved's configuration,
which results in no DNS resolution on the system.

As a workaround, call mount commands directly and remove an empty
/etc/resolv.conf, which can't be blindly removed since users could
have created a static configuration with Ignition.  Also, there is
no touch command in the initramfs, so have bash open the file in
append mode to ensure a mount point exists.